Makes all antags visible to an admin in the orbit menu (#88433)
## About The Pull Request

As the title says, makes all antags visible to an admin in the orbit
menu.

## Why It's Good For The Game

Makes it easier for admins to observe the round and act preventive in
case an antagonist goes too far. A nice QOL addition.


![cooltext471483565189453](https://github.com/user-attachments/assets/c0842cd6-5b4f-4abb-980e-1e7f9d404098)

## Changelog

:cl:
qol: Now all antagonists are visible to an admin in the orbit menu!
/:cl: