Datumizes pAI screen images (#89208)
## About The Pull Request
This PR changes pAI screen images (previously called "emotion icon" or
"card overlay") to be datums.

In doing so, I made some changes in how screen images are handled:
* A screen image's name and icon state are separate now, whereas before
the name determined the icon state used.
* Something useful for downstreams: The icon file that a screen image
pulls from default's to tg's icon file for pAI screen images, but can be
overridden if you don't want to put the new screen image in the same
file.
* The FA icon for each screen image is now selected in the datum, rather
than as part of a constant in some far-off TS file.

Additionally, the codebase previously used inconsistent terminology to
refer to these - "card overlay" in one place, "emotion icon" in another,
"image" somewhere else... Now, all of them are called "screen image" or
some variant thereof. "Screen image" is more accurate to what it is
anyways, since it doesn't need to be a face (although 99% of the time,
it will be).

One last thing: The new datum is in its own directory, all by its
lonesome. If you're wondering why I don't just move it out and put it
next to the rest of the pAI module, it's because I plan to datumize some
other pAI stuff (in particular, software and holochassis) in a future
PR. (Though I'll admit, I'm not sure of the current folder location.)

## Why It's Good For The Game
Datumizes some items that really should've been datums in the first
place.

## Changelog

:cl: MichiRecRoom
refactor: Datumized pAI card screen images
/:cl:

---------

Co-authored-by: SyncIt21 <110812394+SyncIt21@users.noreply.github.com>