Fix jumpsuit sensors context not applying when jumpsuit is held, allow changing sensors when jumpsuit is in active hand (#89391)
## About The Pull Request

I noticed that jumpsuit maxing sensors context tips wouldn't actually
show when it's in your active hand, even though it does work.
Soooo in this pr we simply make it also display when it's the current
held item.

Additionally, we add a `attack_self_secondary(...)` override so the
right-click sensors menu actually works when the jumpsuit is in your
active hand.

This fixes our issues.
## Why It's Good For The Game

It's kinda awkward when it doesn't give you the ctrl-click to max
sensors context when the jumpsuit is in your active hand, even though it
DOES work.
Needing to drop the jumpsuit or swap to the other hand to change sensors
is kinda awkward.
## Changelog
:cl:
fix: Jumpsuit sensors context tips actually show when you are holding
the jumpsuit.
qol: You may right click to change jumpsuit sensors when it's in your
active hand, instead of having to swap to an empty hand or drop it to do
so.
/:cl: